requirements.txt
freud-analysis
MDAnalysis
pandas
pyarrow